+++ tests/delete-requests.spec.js
@@ -23,6 +23,7 @@ describe(`DELETE requests`, () => {
     const unusedResult = await apiClient.deleteFolderByFolderId(folderId);
 
     expect(mockAxios.delete).toHaveBeenCalled();
+    expect(mockAxios.delete).toHaveBeenCalledTimes(1);
     expect(mockAxios.delete).toHaveBeenCalledWith(
       `/folders/4321/`,
       EXPECTED_BASE_AXIOS_CONFIG,
@@ -43,6 +44,7 @@ describe(`DELETE requests`, () => {
     );
 
     expect(mockAxios.delete).toHaveBeenCalled();
+    expect(mockAxios.delete).toHaveBeenCalledTimes(1);
     expect(mockAxios.delete).toHaveBeenCalledWith(
       `/recordings/2468/`,
       EXPECTED_BASE_AXIOS_CONFIG,
@@ -61,6 +63,7 @@ describe(`DELETE requests`, () => {
     const unusedResult = await apiClient.deleteSliceByScorehash(scorehash);
 
     expect(mockAxios.delete).toHaveBeenCalled();
+    expect(mockAxios.delete).toHaveBeenCalledTimes(1);
     expect(mockAxios.delete).toHaveBeenCalledWith(
       `/slices/abcde/`,
       EXPECTED_BASE_AXIOS_CONFIG,
@@ -79,6 +82,7 @@ describe(`DELETE requests`, () => {
     const unusedResult = await apiClient.deleteSliceBySlug(slug);
 
     expect(mockAxios.delete).toHaveBeenCalled();
+    expect(mockAxios.delete).toHaveBeenCalledTimes(1);
     expect(mockAxios.delete).toHaveBeenCalledWith(
       `/scores/1357/`,
       EXPECTED_BASE_AXIOS_CONFIG,