+++ tests/index.spec.js
@@ -23,9 +23,11 @@ describe(`soundslice-data-api`, () => {
         `deleteSliceByScorehash`,
         `deleteSliceBySlug`,
         `duplicateSliceByScorehash`,
+        `getNotationUploadUrlByScorehash`,
         `getRecordingUploadUrlByRecordingId`,
         `getSliceByScorehash`,
         `getSliceBySlug`,
+        `getSliceNotationByScorehash`,
         `getSliceNotationBySlug`,
         `getSliceRecordingsByScorehash`,
         `getSliceRecordingsBySlug`,
@@ -91,6 +93,12 @@ describe(`soundslice-data-api`, () => {
     });
   });
 
+  describe(`getNotationUploadUrlByScorehash`, () => {
+    it(`should be a function`, () => {
+      expect(typeof apiClient.getNotationUploadUrlByScorehash).toBe(`function`);
+    });
+  });
+
   describe(`getRecordingUploadUrlByRecordingId`, () => {
     it(`should be a function`, () => {
       expect(typeof apiClient.getRecordingUploadUrlByRecordingId).toBe(
@@ -114,6 +122,12 @@ describe(`soundslice-data-api`, () => {
   describe(`getSliceNotationBySlug`, () => {
     it(`should be a function`, () => {
       expect(typeof apiClient.getSliceNotationBySlug).toBe(`function`);
+    });
+  });
+
+  describe(`getSliceNotationByScorehash`, () => {
+    it(`should be a function`, () => {
+      expect(typeof apiClient.getSliceNotationByScorehash).toBe(`function`);
     });
   });