+++ tests/post-requests-without-payload.spec.js
@@ -30,9 +30,35 @@ describe(`POST requests without payload`, () => {
     const unusedResult = await apiClient.duplicateSliceByScorehash(scorehash);
 
     expect(mockAxios.post).toHaveBeenCalled();
+    expect(mockAxios.post).toHaveBeenCalledTimes(1);
     expect(mockAxios.post).toHaveBeenCalledWith(
       `/slices/abcde/duplicate/`,
 
+      // the second `axios.post` argument should be `null`
+      // when there is no payload in the request body
+      null,
+
+      EXPECTED_BASE_AXIOS_CONFIG,
+    );
+  });
+
+  test(`.getNotationUploadUrlByScorehash method`, async () => {
+    const mockAxios = makeMockAxios();
+    const apiClient = getApiClientInstance({
+      axios: mockAxios,
+      SOUNDSLICE_APPLICATION_ID,
+      SOUNDSLICE_PASSWORD,
+    });
+
+    const scorehash = `b8vDc`;
+    const unusedResult = await apiClient.getNotationUploadUrlByScorehash(
+      scorehash,
+    );
+
+    expect(mockAxios.post).toHaveBeenCalled();
+    expect(mockAxios.post).toHaveBeenCalledTimes(1);
+    expect(mockAxios.post).toHaveBeenCalledWith(
+      `/slices/b8vDc/notation-file/`,
       // the second `axios.post` argument should be `null`
       // when there is no payload in the request body
       null,
@@ -55,6 +81,7 @@ describe(`POST requests without payload`, () => {
     );
 
     expect(mockAxios.post).toHaveBeenCalled();
+    expect(mockAxios.post).toHaveBeenCalledTimes(1);
     expect(mockAxios.post).toHaveBeenCalledWith(
       `/recordings/2468/media/`,